+++ library/cartridge_get_cached_facts.py
@@ -45,6 +45,7 @@ FACTS_BY_TARGETS = {
         'cartridge_multiversion',
         'cartridge_package_path',
         'cartridge_remove_temporary_files',
+        'cartridge_ignore_split_brain',
         'cartridge_paths_to_keep_on_cleanup',
         'cartridge_run_dir',
         'cartridge_scenario',
@@ -62,6 +63,7 @@ FACTS_BY_TARGETS = {
         'edit_topology_healthy_timeout',
         'edit_topology_allow_missed_instances',
         'expelled',
+        'disabled',
         'failover_priority',
         'instance_discover_buckets_timeout',
         'instance_discover_buckets_retries',
@@ -93,24 +95,32 @@ FACTS_BY_TARGETS = {
         'cartridge_fetch_backups',
         'cartridge_fetch_backups_dir',
     ],
+    'count_disabled_instances': [
+        'instance_info',
+        'disabled',
+    ],
     'facts_for_machines': [
         'expelled',
+        'disabled',
         'ansible_host',
         'ansible_port',
     ],
     'connect_to_membership': [
         'expelled',
+        'disabled',
         'stateboard',
         'config',
     ],
     'alive_not_expelled_instance': [
         'expelled',
+        'disabled',
         'stateboard',
         'config',
         'cartridge_run_dir',
     ],
     'control_instance': [
         'expelled',
+        'disabled',
         'stateboard',
         'config',
         'replicaset_alias',
@@ -118,6 +128,7 @@ FACTS_BY_TARGETS = {
     ],
     'edit_topology': [
         'expelled',
+        'disabled',
         'stateboard',
         'replicaset_alias',
         'roles',
@@ -130,6 +141,7 @@ FACTS_BY_TARGETS = {
     ],
     'failover_promote': [
         'expelled',
+        'disabled',
         'stateboard',
     ],
 }