+++ unit/test_get_cached_facts.py
@@ -26,13 +26,23 @@ class TestGetCachedFacts(unittest.TestCase):
                 'ansible_host': 'some_host',
                 'cartridge_cluster_cookie': 'some-cookie',
                 'cartridge_run_dir': 'some-run-dir',
+                'instance_info': {
+                    'disabled_instances': [],
+                },
             },
             'instance_2': {
                 'stateboard': False,
                 'random_arg': {'test': 'value'},
+                'instance_info': {
+                    'disabled_instances': ['instance_2'],
+                },
             },
             'instance_3': {
+                'disabled': True,
                 'random_arg': {'test': 'value'},
+                'instance_info': {
+                    'disabled_instances': ['instance_2'],
+                },
             },
         })
         self.assertFalse(res.failed, res.msg)
@@ -48,7 +58,27 @@ class TestGetCachedFacts(unittest.TestCase):
                 'instance_2': {
                     'stateboard': False,
                 },
-                'instance_3': {},
+                'instance_3': {
+                    'disabled': True,
+                },
+            },
+            'count_disabled_instances': {
+                'instance_1': {
+                    'instance_info': {
+                        'disabled_instances': [],
+                    },
+                },
+                'instance_2': {
+                    'instance_info': {
+                        'disabled_instances': ['instance_2'],
+                    },
+                },
+                'instance_3': {
+                    'disabled': True,
+                    'instance_info': {
+                        'disabled_instances': ['instance_2'],
+                    },
+                },
             },
             'facts_for_machines': {
                 'instance_1': {
@@ -56,7 +86,9 @@ class TestGetCachedFacts(unittest.TestCase):
                     'ansible_host': 'some_host',
                 },
                 'instance_2': {},
-                'instance_3': {},
+                'instance_3': {
+                    'disabled': True,
+                },
             },
             'connect_to_membership': {
                 'instance_1': {
@@ -66,7 +98,9 @@ class TestGetCachedFacts(unittest.TestCase):
                 'instance_2': {
                     'stateboard': False,
                 },
-                'instance_3': {},
+                'instance_3': {
+                    'disabled': True,
+                },
             },
             'alive_not_expelled_instance': {
                 'instance_1': {
@@ -77,7 +111,9 @@ class TestGetCachedFacts(unittest.TestCase):
                 'instance_2': {
                     'stateboard': False,
                 },
-                'instance_3': {},
+                'instance_3': {
+                    'disabled': True,
+                },
             },
             'control_instance': {
                 'instance_1': {
@@ -88,7 +124,9 @@ class TestGetCachedFacts(unittest.TestCase):
                 'instance_2': {
                     'stateboard': False,
                 },
-                'instance_3': {},
+                'instance_3': {
+                    'disabled': True,
+                },
             },
             'edit_topology': {
                 'instance_1': {
@@ -98,12 +136,16 @@ class TestGetCachedFacts(unittest.TestCase):
                 'instance_2': {
                     'stateboard': False,
                 },
-                'instance_3': {},
+                'instance_3': {
+                    'disabled': True,
+                },
             },
             'failover_promote': {
                 'instance_1': {'expelled': True},
                 'instance_2': {'stateboard': False},
-                'instance_3': {},
+                'instance_3': {
+                    'disabled': True,
+                },
             },
         })