+++ unit/test_get_instance_info.py
@@ -143,6 +143,8 @@ class TestGetInstanceInfo(unittest.TestCase):
                 'some/vinyl/dir/myapp.instance-1',
                 'some/wal/dir/myapp.instance-1',
             ],
+            'disabled_instances': None,
+            'topology_checksum': None,
         })
 
     def test_get_instance_info_multiversion(self):
@@ -195,6 +197,8 @@ class TestGetInstanceInfo(unittest.TestCase):
             'dirs_to_remove_on_cleanup': [
                 'some/data/dir/myapp.instance-1',
             ],
+            'disabled_instances': None,
+            'topology_checksum': None,
         })
 
         # cartridge_package_path isn't specified
@@ -244,6 +248,8 @@ class TestGetInstanceInfo(unittest.TestCase):
             'dirs_to_remove_on_cleanup': [
                 'some/data/dir/myapp.instance-1',
             ],
+            'disabled_instances': None,
+            'topology_checksum': None,
         })
 
     def test_get_stateboard_info(self):
@@ -296,6 +302,8 @@ class TestGetInstanceInfo(unittest.TestCase):
             'dirs_to_remove_on_cleanup': [
                 'some/data/dir/myapp-stateboard',
             ],
+            'disabled_instances': None,
+            'topology_checksum': None,
         })
 
     def test_get_stateboard_info_multiversion(self):
@@ -354,6 +362,8 @@ class TestGetInstanceInfo(unittest.TestCase):
                 'some/vinyl/dir/myapp-stateboard',
                 'some/wal/dir/myapp-stateboard',
             ],
+            'disabled_instances': None,
+            'topology_checksum': None,
         })
 
     def test_paths_to_keep_on_cleanup(self):